Updates to Week 1
These are my answers to the 7 problems. I don't actually call the functions in any of my solutions; my answers are just the functions. Take a look and rip it all apart as much as you like. Also, I commented out the actual problems with comments (docstrings, I think???) Thanks!